+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+#### 1.4.13 (2021-07-05)
+
+##### Bug Fixes
+
+* **kiosk list:** LIV-3289 kiosk sort by serial number fixed (8e6f063d)
+* **Kiosk Details, Organization, Products:** LIV-3446 Components are retouched for Web and Mobile Optimization (263caf44)
+* **product detail:** LIV-3406 product image upload bug fix (7c4e77cf)
+
 #### 1.4.12 (2021-06-25)
 
 ##### Bug Fixes